additional refactoring to account for better naming convention, better spacing/tabs
