+++ api/home/index.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const controller = require('./home.controller')();
+const controller = require('./home-controller')();
 
 router.get('/', controller.home);
