+++ config/express.js
@@ -1,10 +1,10 @@
 const   logger = require('morgan');
-        bodyParser = require('body-parser');
-        cors = require('cors');
+				bodyParser = require('body-parser');
+				cors = require('cors');
 
 
 module.exports = (server, config) => {
-    server.use(logger('dev'));
-    server.use(bodyParser.json());
-    server.use(cors());
+	server.use(logger('dev'));
+	server.use(bodyParser.json());
+	server.use(cors());
 }
