+++ config/routes.js
@@ -3,10 +3,10 @@ const prApi = require('../api/pullrequests');
 const homeApi = require('../api/home');
 
 module.exports = (server) => {
-    // server.use('/users', userApi);
+  // server.use('/users', userApi);
 
-    server.use('/pull-requests', prApi);
-    server.route('*')
-        .get(homeApi);  
+  server.use('/pull-requests', prApi);
+  server.route('*')
+  	.get(homeApi);  
 
 }
